Fix all compiler warnings (-Wsign-compare, -Wunsequenced)
- d_net.cpp: Fix 4 -Wunsequenced UB in multi-byte reads using curByte++
  by sequencing reads into named temporaries before combining
- d_net.cpp: Fix 13 -Wsign-compare warnings: change size_t loop vars to
  int, add (int) casts on MAXPLAYERS comparisons, change 1u literals to 1
- d_netinfo.cpp: Fix 3 -Wsign-compare warnings with (int)MAXPLAYERS casts
- am_map.cpp: Fix 1 -Wsign-compare by casting consoleplayer to unsigned int
- g_game.cpp: Fix 1 -Wsign-compare by casting fr.Read() result to size_t

+++ src/am_map.cpp
@@ -2855,7 +2855,7 @@ void DAutomap::drawPlayers ()
 		}
 
 		// We don't always want to show allies on the automap.
-		if (dmflags2 & DF2_NO_AUTOMAP_ALLIES && i != consoleplayer)
+		if (dmflags2 & DF2_NO_AUTOMAP_ALLIES && i != (unsigned int)consoleplayer)
 			continue;
 		
 		if (deathmatch && !demoplayback &&
